requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-appconf==1.0.5
django-classy-tags==2.0.0
django-cms==3.9.0
django-compressor==2.2
django-formtools==2.3
django-libsass==0.7
django-modeltranslation==0.17.3
django-phone-field==1.8.1
django-sekizai==2.0.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangocms-animate-on-scroll==1.0.1
idna==3.3
libsass==0.21.0
pyTelegramBotAPI==4.1.1
pytz==2021.3
rcssmin==1.0.6
requests==2.26.0
rjsmin==1.0.12
six==1.16.0
soupsieve==2.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
